Simplify tes setup House keeping warnings form compiler
