+++ BookingsApi/BookingsApi.DAL/Helper/UserRoleForHearingRole.cs
@@ -3,7 +3,7 @@
     public static class UserRoleForHearingRole
     {
 
-        public static Dictionary<string, int> UserRoleId = new Dictionary<string, int>
+        public static Dictionary<string, int> UserRoleId => new Dictionary<string, int>
         {
             { UserRoles.Judge, 4 },
             { UserRoles.Individual, 5},
