+++ BookingsApi/BookingsApi.DAL/Mappings/JusticeUserMap.cs
@@ -19,15 +19,4 @@
                 .OnDelete(DeleteBehavior.Cascade);
         }
     }
-
-    public class JusticeUserRoleMap : IEntityTypeConfiguration<JusticeUserRole>
-    {
-        public void Configure(EntityTypeBuilder<JusticeUserRole> builder)
-        {
-            builder.ToTable("JusticeUserRoles");
-            builder.HasKey(x => x.Id);
-            builder.HasOne(x => x.JusticeUser).WithMany(x => x.JusticeUserRoles).HasForeignKey(x => x.JusticeUserId);
-            builder.HasOne(x => x.UserRole).WithMany(x => x.JusticeUserRoles).HasForeignKey(x => x.UserRoleId);
-        }
-    }
 }
