+++ BookingsApi/BookingsApi.Domain/HearingCase.cs
@@ -4,7 +4,6 @@ namespace BookingsApi.Domain
 {
     public class HearingCase : TrackableEntity<long>
     {
-        public long Id { get; set; }
         public long CaseId { get; set; }
         public virtual Case Case { get; set; }
         public Guid HearingId { get; set; }
